requirements.txt
python-dotenv
mongopy
pika
wheel
setuptools